vms.c backslash efs and long name fixes
Message-id: <496552AF.7040702@gmail.com>

This fixes the support for the backslash in a file name as it shows up
in the ExtUtils tests.

Also the flag for rename to work on efs file specifications was
incorrectly the bit number instead of the mask value.

(cherry picked from commit 449de3c2d9bd4f128b1ac6f08de7d29326bc285d)