Restore Win32 build by creating git_version.h.
This solution is based on what is done for config.h (which isn't perfect because it leaves miniperl.exe out-of-sync with the final perl.exe build, but that doesn't really matter).

Also has a problem in that miniperl.exe is used in a couple of places before make_patchnum.pl has been run (to recreate git_version.h and create lib/Config_git.pl), causing the warning:
Warning: failed to load Config_git.pl, something strange about this perl...

The running of make_patchnum.pl also causes some warnings of its own if a ".git" directory is found but "git" isn't on the PATH:
'git' is not recognized as an internal or external command,
operable program or batch file.

These things will be hopefully ironed out later.