Fix handling of DECC$DISABLE_POSIX_ROOT feature on VMS.
The nomenclature is confusing and the CRTL docs even moreso, and
the CRTL implementation seems not to do what the docs say.  The
documented default setting is on, which I take to mean that
processing of SYS$POSIX_ROOT is disabled unless you have asked
for it.  I've followed this behavior even though the CRTL seems
to report the feature off by default.

Also corrected yet another length error in a related call to
simple_trnlnm.