Pathify_dirspec replacement
This replaces pathify_dirspec in vms.c with a new version that better
handles the extended character set.

The [.vms.ext]filespec.t has been adjusted for to support both the
default mode and the extended file spec mode.

This fixes an inconsistency where now vmsify and vmspath will return the
same result for similar input.

Message-ID: <49737F12.6010803@gmail.com>