Add NameAliases.txt; regenerate mktables.list
The NameAliases.txt is an official Unicode database file.  It contains
normative names of Unicode code points that any Unicode implementation
is supposed to recognize.  These are actually corrections to names that
have been published in UnicodeData.txt erroroneously, but to avoid
breaking code that uses the erroneous name, the folks at Unicode create
these additional ones.  (There are about a dozen entries now)

I changed mktables to append these names to Names.pl for use by
charnames.  This might break any existing code that relies on that file
being sorted in numerical code point order.  But if I added them sorted,
it might break code that relied on there being only one entry per code
point.  It's not clear that this file format is considered published and
the api can't change.  The point of the file is to go from the name to
the code point anyway, so the order shouldn't matter to any application
using it for its intended purpose.

mktables.list was not regenerated for 5.1, so was out-of-date.  The
ordering was random (Linux inode order), so the difference listing for
it is large.  I did sort the lists manually and compared them, and they
look ok.  I changed mktables to sort the list, so that this problem
doesn't arise again.

I also corrected some typos in mktables, and added a clarifing comment.