Update Maintainers.pl
I noticed that Maintainers.pl wasn't listing my modules as having the
CPAN version being canonical. The attached patch fixes this.

It also changes the owner of CPANPLUS::Dist::Build to kwilliams
(as per the PAUSE permissions) and removes CPANPLUS::Dist as a
seperate distribution; it always came with CPANPLUS.