Make DB_File use -ldb on MSWin32 with a gcc compiler
DB_File was forcing "-llibdb" for all MSWin32 builds.  That's an MSVC
thing, so this patch makes it use "-ldb" if the compiler is gcc (e.g.
for Strawberry Perl and friends).

Paul Marquess suggested I send a patch to p5p and will update the
source on CPAN when he has a dev box again.

Message-Id: <1232541093-2247-1-git-send-email-dagolden@cpan.org>

Plus: bump $VERSION.