Unbreak build on Windows
Hello p5p,

the recent shakearound of make_patchnum.pl broke the build on Windows,
because make_patchnum.pl lives in the build root, but the Makefile lives
in ./win32, so all targets need to be relative to ./win32.

I've only tested the patch for makefile.mk, as I don't have nmake.exe.
The fix is so simple that it should work on Makefile as well. These are
famous last words.

-max

>From f4d9e594a2a3ea743e9c381d1836a85ada15af85 Mon Sep 17 00:00:00 2001
From: Max Maischein <corion@corion.net>
Date: Sun, 25 Jan 2009 17:34:39 +0100
Subject: [PATCH] Unbreak build on Windows