When loading a PerlIO::via layer, first look in the PerlIO::via namespace
Previously an open "via(Foo)" was first looking for the Foo namespace,
then for PerlIO::via::Foo. That could cause loading problems if both Foo
and PerlIO::via::Foo namespaces existed, and when the latter was
required. The workaround of loading "via(PerlIO::via::Foo)" being rather
ugly, it's preferable to look first for the fully qualified namespace
in order to avoid side-effects depending of what namespaces are
instanciated in the interpreter.