Make t/pod/pod2usage2.t work on case insensitive file systems.
pod_where() checks the current dir when looking for PODs
when the -dirs option is not specified.  Because of
case-insensitivity under Cygwin, this causes t/pod/usage.pod
to get found before lib/Pod/Usage.pm.

The fix is to set -dir=>[] in pod_where() when testing in
CORE under Cygwin.  The attached patch does this.