Skip time64 when there is no 64-bit integer type available.
The y2038 code currently requires that a 64-bit integer type be available,
so the build falls down hard if we try to use it where there isn't one.  A
better fix might be to try NV for time_t and a 32-bit integer for year.

N.B.  We test for the definedness of Quad_t rather than HAS_QUAD since the
latter is explicitly undef'd in the case where 64-bit types are available
but are not being used for Perl's IV and UV.