Remove TODO tests about && in when() clauses
when() actually behaves as documented in perlsyn. It treats "E1 && E2"
expressions as booleans only if E1 is boolean, and if not, it does the
equivalent of the smart match expression "$_ ~~ (E1 && E2)". As long as
the ~~ operator is not expected to distribute boolean expressions found
on its right side, the current behaviour of perl is correct.

So, should we expect ~~ to distribute boolean expressions ? Given that
this would be rather complex to implement and test correctly with
complex expressions, given that we don't expect =~ to distribute either,
and given that when() is already quite smart regarding boolean
operators, I see no motivation for that feature.

This allows to close bug #50538 as "won't fix".