Fix #56194 Regex: (((??{1 + $^N}))) behaves differently in 5.10.0 than in blead
PL_reglastparen and PL_reglastcloseparen contains a pointer are set to & rex->lastparen and & rex->lastcloseparen.
In case END the rex var is modified but PL_reglastparen and PL_reglastcloseparen is not.
Some part of the codes access PL_reglastparen while other parts use rex->lastparen.

This patch corrects this and adds 3 assertions.

I'm currently unable to proof (with a test case) that the code in case EVAL_ab is really nessesary...
Logically speaking it is nessesary but I do not know if it can cause test failures.

Also in the patch are missing regressions between 5.8 -> 5.10 and 5.10 -> 5.11. (and a test script that contains these regressions)

Message-ID: <rt-3.6.HEAD-4802-1236806863-900.56194-15-0@perl.org>

[Includes message and patch edits by committer.]