Make Time::Piece use the init_tm() in util.c, bump to 1.13_03.
Jarkko put it there for that purpose back in e72cf795050cdfe9905e00270c38ba2547626581,
so we seem to be a little slow noticing.

The old broken method for determining whether we need to initialize a tm
struct in the absence of HAS_TM_TM_ZONE being available is preserved for
Perls prior to 5.8.0.