Use files_to_modules() and a CPAN mirror to map modules to bug tracker URLs.
Output two hashes, ready for inclusion in Module::CoreList. %upstream maps
module names to the UPSTREAM status in Porting/Maintainers.pl. %bug_tracker
maps module names to the URL of the bug tracker, or undef if explict bug
tracker is not known for a module known to be dual life.