Document that XS_VERSION_BOOTCHECK may break if $VERSION is a long enough NV
As seen in <http://bugs.debian.org/482139>, using a long floating point
number as $VERSION can make XS_VERSION_BOOTCHECK fail because of a loss
of precision in the stringification:

RRDs object version 1.2999070803 does not match bootstrap parameter 1.29990708 at /usr/lib/perl/5.10/DynaLoader.pm line 219.

The mismatch happens in Perl_upg_version(), which converts a double value
(NV) to a string with nine decimals but leaves a string value intact.

Using a string as $VERSION works fine:

-$VERSION=1.299907080300;
+$VERSION="1.299907080300";

so recommend this in the VERSIONCHECK entry of perlxs.pod.