Export two more PerlIO symbols.
Ben Morrow noticed that:

PerlIO_push is exported, but PerlIO_find_layer is not, so getting a
PerlIO_funcs* to pass to it is difficult.

Several exported functions (PerlIO_parse_layers, PerlIO_layer_fetch,
PerlIO_list_free, ...) take a PerlIO_list_t* argument, but
PerlIO_list_alloc is not exported, so these are effectively useless.