Merge the bulk of Perl_magic_clearsig() into Perl_magic_setsig().
The comment at the top of Perl_magic_clearsig() said:
    XXX Some of this code was copied from Perl_magic_setsig. A little
    refactoring might be in order.
and it was not wrong.

Perl_magic_clearsig() is almost equivalent to Perl_magic_setsig() with the new
signal handler as "DEFAULT". Externally, the sv parameter to Perl_magic_setsig()
was Not NULL, so use a NULL sv to signal that the code is being called as
Perl_magic_clearsig(), for the places where the behaviour of the two diverges.