There is no pthread_sigmask on VMS, so use sigprocmask instead.
According to the standard, use of sigprocmask is undefined in a
multi-threaded context, but in the absence of a per-thread signal
blocking mechanism (likely impossible until and unless per-thread
ASTs become available), this may be the best we can do.

This is a follow-up to: b762d8667351cb765bc1d6419d30acff085da502.