simplify PERL_MEM_LOG
This combines multiple environment variable reads into 1,
where it looks for values like "2mst"
-2 leading digits are atoi()d to get FD
-m memory logging please
-s sv logging also
-t timestamp those please.

Combining these reduces overhead such that it seemed
worthwhile to drop all the ifdefs.  TBD whether this works
in the environment that drove the original tradeoffs.

If it isnt enough, Id be tempted by a global static ptr,
and on 1st use, is read, seen 0, a lock is taken, and getenvar
run to populate it, unlocked, proceed.  This would remove
iterative overheads.