Add security contact information to perlsec
This patch inserts a short paragraph with security contact
information near the top of the "Perl Security" documentation
page.  This would seem a likely place someone would look for
such information (rather than INSTALL or perldelta where it
lives today).  I've put it at the top, not the bottom
to make it easier to find.