cleanup the EUI/EUMM test mess, and bump Install.pm to 1.54
This removes lib/ExtUtils/t/basic.pl and t/basic_finish.pl, and
replaces them with the EUI specific t/InstallWithMM.t. The
version bump is purely to allow the release the new version on
CPAN and to make it possible to include a non dev release in
Perl 5.10.1 if Dave wants to.

Note that while there are no real code changes in this commit I
did trim whitespace from Install.pm, so if you want to see the
real diff here please use diff -w.