bump-perl-version false positives
In the runup to 5.10.1, Dave created bump-perl-version to automate
the tedious process of incrementing the Perl version number.  Trying
it out on blead, I found that there are a few files it matches
which either mention a version of Perl and then say "replace 5.x.y with
your version of Perl" or which are describing changes in a particular
version of perl. The attached patch skips those files.

At this point, we're down to 18 files which need to be changed to bump
the version at release time.

-j

From 0c347ba41fea91bbe36b53841cc150002c8782d5 Mon Sep 17 00:00:00 2001
From: Jesse Vincent <jesse@bestpractical.com>
Date: Wed, 29 Jul 2009 10:00:09 -0400
Subject: [PATCH] Removed some false positives flagged for gratuitious changing by the bump-perl-version releng tool