Add support for Abstract namespace sockets
Abstract namespace sockets are Linux-specific socket type that live in
AF_UNIX family, slightly abusing it to be able to use arbitrary
character arrays as addresses: They start with nul byte and are not
terminated by nul byte, but with the length passed to the socket()
system call.

Added regression test for the correct address length computation.

Signed-off-by: Lubomir Rintel <lkundrak@fedoraproject.org>