Share common code in t/TEST and t/harness, by having harness require ./TEST
The logical way to do this would be to have the common code in a file both
require or use. However, t/TEST needs to still work, to generate test results,
even if require isn't working, so we cannot do that. t/harness has no such
restriction, so it is quite acceptable to have it require t/TEST.