Use & function-call notation to disable prototype checking on dl_load_file()
This causes the test for insufficient arguments to work on Win32 too, where
(for some reason) DynaLoader.xs is built *with* prototypes on the XS code.
On *nix, where it is build by MakeMaker in ext/DynaLoader/, it is built without
prototypes on the XS code, so the check on the number of parameters for
dl_load_file() is purely run-time, which is what the test script used to assume.