set PERL_LEGACY_UNICODE_CHARCLASS_MAPPINGS to 0 and enable proper POSIX char class matching
This also alters which Unicode properties that the POSIX character
class and the Perl "special" character classes, like \w and \d map
to. At the same time it allows a number of tests for POSIX character
class behaviour to be switched from todo to non todo. Legacy testing
is still available by changing the define and setting the
PERL_TEST_LEGACY_POSIX_CC value to true.