OS/2 hadn't been updated to cope with the ext/ restructuring.
I don't have OS/2, so I can't test this, but the code in Configure will assume
flat directories, because ext/File-Glob is present, and hence not search
recursively and not find the OS/2 extensions if they are copied into ext/OS2/*

I believe that without this change OS/2 will not have been building since the
change to flattened ext. This change may not be sufficient to get OS/2
building again, but it is in the right direction.