win32/FindExt.pm thought that ext/List-Util was a nonxs extension.
(It needed a change analogous to 75cdbdc152433e4b - prior to that, Configure
was looking for ext/$dir/$leaf.xs, which was no longer a valid test after the
upgrade to Scalar-List-Utils 1.21 (2ff286160a117d4e), which renamed Utils.xs
to ListUtil.xs)

win32/FindExt.pm still considers IO-Compress as a nonxs extension.
Configure was special-cased to treat this as dynamic (ie architecture
specific) in 4cc80fc4cbbd82e2.