Always load the re.xs code, unless we're running under miniperl.
This is a lot simpler than the previous implementation, which would lazy load
the XS code only if it was needed, and then jump through more hopes to try to
give the user a reasonable error message if it failed.

In the real world, people either have a full perl installed with the re
extension, or their install is already horribly broken. Don't let code related
to bootstrapping the core build pollute the installation.