No need to run AutoSplit on lib - all AutoLoader using modules are in ext.
It seems that the code in the Makefile to run AutoSplit over lib has actually
been vestigial since some point *before* 5.000. Presumably, initially, Perl code
for extensions was in lib, and autosplit (the precursor to AutoSplit.pm) was run
on them there. Or they were copied from ext to lib first, before splitting.
Fifteen year-old backup tapes refuse to give us the detailed history of the
pre-5.000 period.

Whatever the cause, the code was never needed - backed up by the absense of any
analgous code in the Win32 Makefiles. So remove it. This happens to also remove
the only use of the shell script makedir, so remove it too.