Reuinte File::Spec with the rest of the PathTools distribution in ext.
It can't really be renamed from ext/Cwd to ext/PathTools, because Configure and
Makefile.SH need to know the name of the shared object produced, and they infer
this from the name of the directory.

This concludes the migration of modules from lib to ext. Exporter and version
remain in lib, but I don't think that there is any benefit in trying to move
either of them.