Fix when( scalar ... ) bug
looks_like_bool should look past scalar(). Prior to this fix, saying
scalar() caused the when() argument to be /not/ a boolean even if it
reasonably was. Perhaps something like: when( scalar /$re/ ). I don't
know any natural way someone is doing to use when(scalar ...) but it
certainly seems like it was mistakenly left out.