git config lossage
The git that I have installed (Debian etch) doesn't support "git
config".
This causes nasty failure in make_patchnum.pl, where it gets a help
message from git, and then blindly interpolates that message into a
command that it passes to the shell for execution.  The attached patch
fixes this.