in regexec.c move the BOUND logic out of the way of the special CC logic
This is a first step towards macroizing the special CC handler logic so
it is easier to maintain them, for instance interestng optimisations are
being used in one, but not all, even though the logic is sharable. By
moving the BOUND logic out of the way the code repition is much clearer.