Move the require './test.pl' to the end of t/comp/hints.t
Ideally tests in t/comp wouldn't use require, as require isn't tested yet, but
this test really needs runperl(), and really wants to live in t/comp/hints.t,
so place it at the end, so that any catestrophic failure only fails the last
test. We don't use any other functionality of t/test.pl

This test uses hard-coded test numbers, but I'm not convinced that it would be
correct to re-write it to use an automatically incrementing counter, as that
wouldn't fail in an obvious fashion if some compile-time blocks ran out of
order. What we have *will* fail in an informative fashion if compile time
blocks do not run correctly.