Exclude CGI/t/fast.t from the core, as it relies on FCGI, which is not in core.
Whilst the test is designed to skip if FCGI is not present, that logic is now
broken, and even if it were fixed, the as-implemented skip count is now wrong.
We're never going to run this test, as core tests run with just the core's
libraries in @INC, so I see no reason to ship it.