MakeMaker::Test::Utils::perl_lib now copes with relative paths for core testing.
In the core, @INC already contains the moral equivalent of blib/lib. However,
it's a relative path (by default), so make it absolute. It's easier to KISS if
this is done *before* any change of directory, so document this, and change the
non-core case to add the absolute path of 'blib/lib' to @INC, rather than the
absolute path of '../blib/lib'.