Remove Test-Simple's 00compile.t from core
It required a hand-edited version of Test-Simple's MANIFEST due to the
exclusion of a .pm file from core, but we shouldn't have edited versions
of CPAN files in core. The presence of the MANIFEST file also caused
warnings from the Makefile.PL due to other files being excluded from core,
so rather than further editing of the MANIFEST it is simplest to just
remove it and the test that requires it.