perlvar $1 clarification
$1 is currently documented as being set by the "last pattern match".
But it is left alone by unsuccessful pattern match attempts (continuing
to hold a value from an earlier successful match).

Saying "last successful pattern match" clarifies this; it's also the
phrase used to document $&.

Second, the entry for $1 in perlvar doesn't actually contain the text
"$1" anywhere.  As such, doing man perlvar then using /\$1 to search for
it in less doesn't locate it (though does match other places in that
file where $1 happens to be used).