Remove the entry about removing Perl_deprecate()
As a minor C function, I feel that at most it should be in "changed internals".
However, as it was not part of the public API, and searches showed that nothing
was using it, likely nothing is going to break if it goes. And if any naughty
code does break:
a: It will be at link time
b: It will be easy to find the cause from git log

I also feel that it's worth removing it, as existing perldeltas are the most
obvious guide for new volunteers in terms of "write something like this".
So having bad examples isn't going to help educate future authors.