Porting/makerel no longer massages win32 files during the final tarball build process
I _believe_ that this change will mean that our shipped tarball will
contain all files exactly as they are in the repo - The only difference
will be in file permissions.

I'm told that git can version the execute bit on files. If I'm feeling
especially adventurous, I may write a porting test to ensure that the
right files (and only the right files?) are marked +x in git and then
remove that bit from Porting/makerel. I wouldn't mind it if someone beat
me to that.