Add mingw-w64 non-crosscompiling config files
Files provided by Sisyphus with this comment: "Attached are the 2
addtional files that we need in the win32 folder. (I did a dos2unix on
them so, hopefully the line endings will be right.) These files are
needed to accommodate the particular 64-bit mingw compiler that
Strawberry Perl will be using. They'll actually be used as the default
64-bit mingw templates."